[#22] refactor: final 사용 및 decimal 사용
- 점수 계산에 적당한 scale인 5,2를 넣었습니다.

+++ jaknaeso-core/src/main/resources/sql/022/004_create_table_keyword_scores.sql
@@ -1,6 +1,6 @@
 create table if not exists keyword_scores
 (
-    score            double,
+    score            decimal(5,2),
     survey_option_id bigint not null,
     keyword          varchar(30)
 );
